scripts/requirements.txt
pandas
pyproj
psycopg2-binary
requests
python-dotenv
osmium
